Refactor: 파일 업로드 설정 파일 이동 및 변경
