+++ src/main/java/com/filmdoms/community/account/config/AmazonS3Config.java
@@ -1,4 +1,4 @@
-package com.filmdoms.community.imagefile.amazons3;
+package com.filmdoms.community.account.config;
 
 import com.amazonaws.auth.AWSCredentials;
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
@@ -8,9 +8,11 @@ import com.amazonaws.services.s3.AmazonS3ClientBuilder;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.PropertySource;
 
 
 @Configuration
+@PropertySource(value = "classpath:s3.yml", factory = YamlPropertySourceFactory.class)
 public class AmazonS3Config {
 
     @Value("${cloud.aws.credentials.accessKey}")
